requirements.txt
requests
bs4
fastapi
lxml
tk
pillow
customtkinter
